First part of generic point cloud writing

+++ src/packet.rs
@@ -70,7 +70,7 @@ pub struct DataPacketHeader {
 }
 
 impl DataPacketHeader {
-    pub const SIZE: u64 = 6;
+    pub const SIZE: usize = 6;
 
     pub fn read(reader: &mut dyn Read) -> Result<Self> {
         let mut buffer = [0_u8; 5];
@@ -104,7 +104,7 @@ impl DataPacketHeader {
     }
 
     pub fn write(&self, writer: &mut dyn Write) -> Result<()> {
-        let mut buffer = [0_u8; Self::SIZE as usize];
+        let mut buffer = [0_u8; Self::SIZE];
         buffer[0] = 1;
         let flags = if self.comp_restart_flag { 1_u8 } else { 0_u8 };
         buffer[1] = flags;
